LatLon.jl Documentation

Overview

This module computes geo-distances between points and between sets. Also provided are retrieve functions for geo-data from NASA via KML files.

The distance functions can be applied to the data from the geo-data from NASA files. However, it should be noted that the distance functions assume a perfect sphere and do not, for instance, correct for the polar flattening of most planetary objects.

NOTE: This module uses a constant internal structure, MC. The values from this structure are used in several of the functions below. In addition, some functions use values from this structure as default values for some of the arguments.

Exports

Geo-Distance Functions:

geo_dist

Computes the distance between two points on a sphere. The default sphere is the earth with the radius given in Kilometers.

geo_midpoint

Computes the mid lat/lon between to lat/lon coordinates.

latlon_set_dist

Computes the distance between two geo-sets on a sphere.

Geo-Extraction Functions:

center_latlon_from_NASA_xml_file

Retrieve lat/lon data from a NASA XML "center" file.

upath_latlon_from_NASA_xml_file

Retrieve lat/lon data from a NASA XML "upath" file.

Geo-Distance Functions

LatLon.geo_dist — Function

geo_dist(coord1::Vector{Float64}  , 
         coord2::Vector{Float64}  ,
         radius=MC.radius::Float64 )

Computes the distance between two coordinates on a sphere represented as two lon/lat vectors in signed degrees. That is, north latitude is positive, south latitude is negative, while east longitude is positive and west longitude is negative. The radius, R, defaults to the Earth's radius in Kilometers.

Details

Dot product, dp, of the geo-positions on the unit sphere gives the cosine of the angle between the two points (on the "great circle"). This is the information contained in coord1 and coord2. Once the cosine is computed we can easily retrieve the angle (in Radians) and then it is easy to find the distance between the two points – the length of the arc connecting them on a "great circle".

The Cartesian points of the lon/lat on the unit sphere are:

v1 = ( cos(θ1)cos(ϕ1), sin(θ1)cos(ϕ1), sin(ϕ1) )

v2 = ( cos(θ2)cos(ϕ2), sin(θ2)cos(ϕ2), sin(ϕ2) )

The dot product becomes:

dp = v1 ⋅ v2 = cos(ϕ1)cos(ϕ2) ( cos(θ1)cos(θ2) + sin(θ1)sin(θ2) ) + sin(ϕ1)sin(ϕ2)

Simplifying

dp = cos(ϕ1)cos(ϕ2) cos(θ1 - θ2) + sin(ϕ1)sin(ϕ2)

dp = ( cos(ϕ1)cos(ϕ2) + sin(ϕ1)sin(ϕ2) ) cos(θ1 - θ2) + (1 - cos(θ1 - θ2)) sin(ϕ1)sin(ϕ2)

dp = cos(θ1 - θ2) cos(ϕ1 - ϕ2) + (1 - cos(θ1 - θ2)) sin(ϕ1)sin(ϕ2)

Set A = cos(θ1 - θ2), B = cos(ϕ1 - ϕ2), then

dp = A B + (1 - A) sin(ϕ1) sin(ϕ2)

The angle between vectors in Radians is then ψ = acos(dp)

Distance between the two points on the "great circle", Δ, is

Δ = R ψ

Arguments

coord1::Vector{Float64} – A 2-element vector: [lon, lat] in signed degrees.

coord2::Vector{Float64} – A 2-element vector: [lon, lat] in signed degrees.

R::Float64 – The radius of the sphere (Default is Earth's radius in Kilometers.)

Return

The distance in (in the units of the radius, R) via a "great circle" path.

LatLon.geo_midpoint — Function

geo_midpoint(coord1::Vector{Float{64}, 
             coord2::Vector{Float64}  )

Computes and returns the mid-point of two points on the sphere as a vector (lon/lat) in signed decimal degrees.

Arguments

coord1::Vector{Float64} - A 2-element vector: [lon, lat] in signed degrees.

coord2::Vector{Float64} - A 2-element vector: [lon, lat] in signed degrees.

Return

The 2-Vector representing the longitude and latitude.

LatLon.latlon_set_dist — Function

latlon_set_dist(coord1s::Vector{Float64}, 
                coord2s::Vector{Float64},
                R=MC.radius::Float64     )

Computes the set distance between two sets as represented by their lat/lon coordinates, in singed decimal degrees. It does this the hard way by computing the distance of all pairs of points between the two sets. The radius, R, defaults to the Earth's radius in Kilometers.

Arguments

coord1s::Matrix{Float64} – The coordinates of the first set.

coord2s::Matrix{Float64} – The coordinates of the second set.

R::Float64 – Radius of sphere (Default is Earth's radius in Kilometers.)

Return

The minimum distance (in the same units as R) between the sets along with the index of the points for each set representing the closest points from each set.

A Tuple: (distinkm, set1index, set2index)

Geo-Extraction Functions

LatLon.center_latlon_from_NASA_xml_file — Function

center_latlon_from_NASA_xml_file(file::String)

Retrieves the latitude and longitude from a "center" NASA kml file as a matrix of Float64: 2xN. Here N is the number of points. Each column of this matrix (a 2-vector) is a lon/lat pair, in signed decimal degrees.

LatLon.upath_latlon_from_NASA_xml_file — Function

upath_latlon_from_NASA_xml_file(file::String)

Retrieves the latitude and longitude of the annular ring representing the boundary of the "total" part of an eclipse from a "upath" NASA kml file as a matrix of Float64: 2xN. Here, N is the number of points. Each column of this matrix (a 2-vector) is a lon/lat pair, in signed decimal degrees.

Index

LatLon.center_latlon_from_NASA_xml_file

LatLon.geo_dist

LatLon.geo_midpoint

LatLon.latlon_set_dist

LatLon.upath_latlon_from_NASA_xml_file